Merge Deliverable 4.3 final

+++ lib_archi/base_application_service.py
@@ -1,6 +1,7 @@
+import json
+from lib_archi.abstractions.request_interface import IRequest
 from .base_repository import BaseRepository
-from typing import TypeVar, Generic, Optional, List, Dict
-
+from typing import Any, TypeVar, Generic, Optional, List, Dict
 
 Entity = TypeVar('Entity')
 
@@ -17,7 +18,7 @@ class BaseApplicationService(Generic[Entity]):
             for CRUD operations on entities.
     """
 
-    def __init__(self, repository: BaseRepository[Entity]):
+    def __init__(self, repository: BaseRepository[Entity], logic_map: Dict[str, Any] = None):
         """Initializes the service with a repository instance.
 
         Args:
@@ -25,8 +26,20 @@ class BaseApplicationService(Generic[Entity]):
                 entity operations.
         """
         self.repository = repository
+        self.logic_map = logic_map or {}
 
-    def get(self, ids: Dict) -> Entity:
+    def inject_logic(self, verb: str ) -> Optional[Any]:
+        try:
+            logic = self.logic_map.get(verb)
+            if logic:
+                return logic
+            else:
+                return None
+            
+        except Exception as e:
+            return None
+
+    def get(self, request: IRequest) -> Entity:
         """Retrieves an entity by its unique identifier.
 
         Args:
@@ -35,9 +48,14 @@ class BaseApplicationService(Generic[Entity]):
         Returns:
             Entity: The entity retrieved from the repository.
         """
-        return self.repository.get(ids)
+        logic = self.inject_logic("get")
+        if logic:
+            return logic(request, self.repository)
+        else:
+            ids = request.get_path_params()
+            return self.repository.get(ids)
 
-    def get_collection(self, ids: Dict) -> List[Entity]:
+    def get_collection(self, request: IRequest) -> List[Entity]:
         """Retrieves a collection of all entities.
 
         Returns:
@@ -50,9 +68,25 @@ class BaseApplicationService(Generic[Entity]):
         else:
             <something>
         """
-        return self.repository.get_collection(ids)
+        logic = self.inject_logic("get_collection")
+        
+        if logic:
+            return logic(request, self.repository)
+        else:
+            ids = request.get_path_params()
 
-    def post(self, entity: Entity, ids: Dict) -> Optional[Entity]:
+            query = request.get_query_params()
+            query = query.get('filter', {}) if isinstance(query, dict) else {}
+    
+            try:
+                query = json.loads(query)
+            except (json.JSONDecodeError, TypeError):
+                query = {}
+            
+            query = {**ids, **query}
+            return self.repository.get_collection(query)
+
+    def post(self, entity: Entity, request: IRequest) -> Optional[Entity]:
         """Creates a new entity in the repository.
 
         Args:
@@ -61,9 +95,14 @@ class BaseApplicationService(Generic[Entity]):
         Returns:
             Optional[Entity]: The created entity, or None if creation fails.
         """
-        return self.repository.post(entity, ids)
+        logic = self.inject_logic("post")
+        if logic:
+            return logic(request, self.repository, entity)
+        else:
+            ids = request.get_path_params()
+            return self.repository.post(entity, ids)
 
-    def put(self, entity: Entity, ids: Dict) -> Optional[Entity]:
+    def put(self, entity: Entity, request: IRequest) -> Optional[Entity]:
         """Fully updates an existing entity in the repository.
 
         Args:
@@ -72,9 +111,14 @@ class BaseApplicationService(Generic[Entity]):
         Returns:
             Optional[Entity]: The updated entity, or None if the update fails.
         """
-        return self.repository.put(entity, ids)
+        logic = self.inject_logic("put")
+        if logic:
+            return logic(request, self.repository, entity)
+        else:
+            ids = request.get_path_params()
+            return self.repository.put(entity, ids)
 
-    def patch(self, entity: Entity, ids: Dict) -> Optional[Entity]:
+    def patch(self, entity: Entity, request: IRequest) -> Optional[Entity]:
         """Partially updates an existing entity in the repository.
 
         Args:
@@ -83,9 +127,14 @@ class BaseApplicationService(Generic[Entity]):
         Returns:
             Optional[Entity]: The updated entity, or None if the update fails.
         """
-        return self.repository.patch(entity, ids)
+        logic = self.inject_logic("patch")
+        if logic:
+            return logic(request, self.repository, entity)
+        else:
+           ids = request.get_path_params()
+           return self.repository.patch(entity, ids)
 
-    def delete(self, ids: Dict) -> Optional[Entity]:
+    def delete(self, request: IRequest) -> Optional[Entity]:
         """Deletes an entity by its unique identifier.
 
         Args:
@@ -94,4 +143,9 @@ class BaseApplicationService(Generic[Entity]):
         Returns:
             Optional[Entity]: The deleted entity, or None if deletion fails.
         """
-        return self.repository.delete(ids)
+        logic = self.inject_logic("delete")
+        if logic:
+            return logic(request, self.repository)
+        else:
+           ids = request.get_path_params()
+           return self.repository.delete(ids)
